Renamed OrderMap -> IdentifiedVecOf, split it into many files, it is not longer Deref for IndexMap, add more tests. Rename 'get' to 'get_id' (unsure about this).
