+++ src/core/types/identified_vec_of/mod.rs
@@ -1,0 +1,23 @@
+mod identifiable;
+mod identified_vec_of;
+mod identified_vec_of_display_debug;
+mod identified_vec_of_iterator;
+mod identified_vec_of_modify;
+mod identified_vec_of_query;
+mod identified_vec_of_serde;
+mod identified_vec_of_uniffi_converter;
+
+#[cfg(test)]
+mod user;
+
+#[cfg(test)]
+use user::*;
+
+pub use identifiable::*;
+pub use identified_vec_of::*;
+pub use identified_vec_of_display_debug::*;
+pub use identified_vec_of_iterator::*;
+pub use identified_vec_of_modify::*;
+pub use identified_vec_of_query::*;
+pub use identified_vec_of_serde::*;
+pub use identified_vec_of_uniffi_converter::*;
