+++ src/core/types/mod.rs
@@ -7,13 +7,13 @@ mod decimal192_uniffi_fn;
 mod entity_kind;
 mod epoch;
 mod exactly_n_bytes;
+mod identified_vec_of;
 mod keys;
 mod locale_config;
 mod logged_result;
 mod non_empty_max_n_bytes;
 mod nonce;
 mod nonce_uniffi_fn;
-mod ordered_map;
 mod requested_number_quantifier;
 mod requested_quantity;
 mod requested_quantity_uniffi_fn;
@@ -31,13 +31,13 @@ pub use decimal192_uniffi_fn::*;
 pub use entity_kind::*;
 pub use epoch::*;
 pub use exactly_n_bytes::*;
+pub use identified_vec_of::*;
 pub use keys::*;
 pub use locale_config::*;
 pub use logged_result::*;
 pub use non_empty_max_n_bytes::*;
 pub use nonce::*;
 pub use nonce_uniffi_fn::*;
-pub use ordered_map::*;
 pub use requested_number_quantifier::*;
 pub use requested_quantity::*;
 pub use requested_quantity_uniffi_fn::*;
