+++ src/lib.rs
@@ -39,7 +39,9 @@ pub mod prelude {
     pub use radix_engine_common::math::traits::CheckedMul as ScryptoCheckedMul;
     pub(crate) use std::cmp::Ordering;
     pub(crate) use std::collections::BTreeMap;
+    pub(crate) use std::fmt::{Debug, Display, Formatter};
     pub(crate) use std::fs;
+    pub(crate) use std::hash::Hash as StdHash;
     pub use std::ops::{Add, AddAssign, Deref, Div, Mul, Neg, Sub};
     pub(crate) use std::str::FromStr;
     pub(crate) use std::sync::Arc;
