+++ src/profile/v100/app_preferences/security.rs
@@ -48,7 +48,7 @@ pub struct Security {
     pub is_cloud_profile_sync_enabled: bool,
     pub is_developer_mode_enabled: bool,
     pub structure_configuration_references:
-        OrderedMap<SecurityStructureConfigurationReference>,
+        IdentifiedVecOf<SecurityStructureConfigurationReference>,
 }
 
 impl Security {
@@ -56,7 +56,7 @@ impl Security {
     pub fn new(
         is_cloud_profile_sync_enabled: bool,
         is_developer_mode_enabled: bool,
-        structure_configuration_references: OrderedMap<
+        structure_configuration_references: IdentifiedVecOf<
             SecurityStructureConfigurationReference,
         >,
     ) -> Self {
@@ -71,19 +71,19 @@ impl Security {
 impl Default for Security {
     /// By default we cloud profile sync is enabled and developer mode is disabled, with an empty `structure_configuration_references` list.
     fn default() -> Self {
-        Self::new(true, false, OrderedMap::new())
+        Self::new(true, false, IdentifiedVecOf::new())
     }
 }
 
 impl HasSampleValues for Security {
     /// A sample used to facilitate unit tests.
     fn sample() -> Self {
-        Self::new(true, true, OrderedMap::new())
+        Self::new(true, true, IdentifiedVecOf::new())
     }
 
     /// A sample used to facilitate unit tests.
     fn sample_other() -> Self {
-        Self::new(false, false, OrderedMap::new())
+        Self::new(false, false, IdentifiedVecOf::new())
     }
 }
