+++ src/wrapped_radix_engine_toolkit/high_level/manifest_building/third_party_deposit_update/third_party_deposits_delta.rs
@@ -38,7 +38,7 @@ impl ThirdPartyDepositsDelta {
                     !to.assets_exception_list
                         .clone()
                         .unwrap_or_default()
-                        .contains(x)
+                        .contains_by_id(x)
                 })
                 .map(ScryptoManifestValue::from)
                 .collect(),
